requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
aiofiles==23.2.1
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3
tqdm==4.66.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
scikit-learn==1.3.2